requirements.txt
celery==5.2.2
python-decouple==3.1
Django==3.2.14
channels==3.0.3
django-cors-headers==3.5.0
requests==2.24.0
xlrd==1.2.0
aiohttp==3.7.4
lxml==4.9.1
xlsxwriter==1.3.7
redis==3.5.3
channels-redis
elasticsearch==7.10.0
pandas==1.1.4
psycopg2-binary
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
flower==0.9.5
uvicorn==0.14.0
websockets==10.3
graphql-ws==0.4.4
graphene-django==2.15.0